requirements.txt
requests
discord
python-dotenv
selenium
async